[03663] Extract ISessionParser interface and implement parsers
