+++ src/Ivy.Tendril.Test/ModelPricingServiceTests.cs
@@ -1,4 +1,5 @@
 using Ivy.Tendril.Services;
+using Ivy.Tendril.Services.SessionParsers;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 
@@ -6,7 +7,9 @@ namespace Ivy.Tendril.Test;
 
 public class ModelPricingServiceTests
 {
-    private readonly ModelPricingService _service = new(NullLogger<ModelPricingService>.Instance);
+    private readonly ModelPricingService _service = new(
+        NullLogger<ModelPricingService>.Instance,
+        new ISessionParser[] { new ClaudeSessionParser(), new CodexSessionParser(), new GeminiSessionParser() });
 
     [Fact]
     public void LoadsEmbeddedModelsYaml()
@@ -429,7 +432,8 @@ public class ModelPricingServiceTests
     public void GetPricing_LogsWarning_WhenModelNotFound()
     {
         var logger = new CapturingLogger<ModelPricingService>();
-        var service = new ModelPricingService(logger);
+        var service = new ModelPricingService(logger,
+            new ISessionParser[] { new ClaudeSessionParser(), new CodexSessionParser(), new GeminiSessionParser() });
 
         service.GetPricing("totally-unknown-model");
 
@@ -443,7 +447,8 @@ public class ModelPricingServiceTests
     public void GetPricing_DoesNotLogWarning_WhenModelFound()
     {
         var logger = new CapturingLogger<ModelPricingService>();
-        var service = new ModelPricingService(logger);
+        var service = new ModelPricingService(logger,
+            new ISessionParser[] { new ClaudeSessionParser(), new CodexSessionParser(), new GeminiSessionParser() });
 
         service.GetPricing("claude-opus-4");
