+++ src/Ivy.Tendril/Services/SessionParsers/ISessionParser.cs
@@ -1,0 +1,7 @@
+namespace Ivy.Tendril.Services.SessionParsers;
+
+public interface ISessionParser
+{
+    string Name { get; }
+    CostCalculation Parse(string filePath, IModelPricingService pricingService);
+}
